Reorganize files for clarity
