+++ src/components/Card/Card.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import './Card.css';
 
 const Card = ( {image, title, releaseDate, voteAverage, overview} ) => {
